Create fetch_qt_version.py
This adds a python script/module that will return or print the Qt version of the Qt codebase provided, given a path to the base of the `qt5` base repository.
This requires the qtbase module to be checked out, and will return the version of qtbase. It is possible that other modules are checked out to different Qt versions, however this just takes into account qtbase - which is a good sign of what overall Qt version the codebase is.