Fix when comments/other lines reference QT_REPO_MODULE_VERSION
If there are other lines other than the `set(QT_REPO_MODULE_VERSION` line that have `QT_REPO_MODULE_VERSION`, this script fails with "qtbase doesn't exist" even if it does exist - because an `IndexError: 'list index out of range'` exception is thrown when trying to index [1] of the string split on ". This happens when it processes a line with `QT_REPO_MODULE_VERSION` that does not contain a double quote in it.