Add python tool to extract real estate tax pdf page
Town of York, Maine provides electronic copies of real estate taxes, but they are in the form of a single enormous PDF with one page per parcel in the entire town of York - as of 2023, about 11,500 pages. This tool takes that PDF, along with a last name search query, and will extract the first matching PDF page that matches the last name given.