Reduce animated gif by removing odd frames
This adds a tool that will reduce the size of animated gifs by removing
every other frame. It does assume there is constant delay between
frames, and that the frames are not alpha ones that have selective
built-in erasing of the prior frame's content where not the same.

This is where it didn't work for the case I wrote it for, and I didn't
get it to work properly with that gif. That gif is this one, which is
added as part of this commit, in case it disappears.
https://www.spaceweatherlive.com/images/help/zonnewind/L1_animation.gif

So - note this script will indeed remove every other frame, but you may
need to fix it to do things perfectly.

Assumptions:
  * main assumption above (that there is no dependency on prior frames)
  * Doesn't check to see if the gif loops infinitely or finitely, and
    leaves the default, which is to loop infinitely.