Handle when scale factor line has more than floatx
The script fails when the scale factor line contains more than just a
line that starts with a float number and ends with a single non-float
character. This alters the logic to parse the float to pull out the
float characters with a regular expression first, then do the float
conversion.